Bunch of changes: -Renamed older wav player python files ver1OLDER -Added "final version" of wav_player_ver_16_WORKS -Added first version of Audio_Menu_ver1
wav_player_ver_16_works--is the "final" version of the working BASIC RaspPi song player.  This version only uses ONE sensor switch, and the rewind button.  Works fine, needs additional features added in the next version.

New Audio_menu is to use the extra toggle switch and allow the user to select which songs or which mode to use for the box.